feat: add fix for userid change
